Extract shared koffi mock helper for GciLibrary tests
